refactor searchbar
