+++ src/App.js
@@ -1,7 +1,7 @@
 import JobApplicationList from './components/JobApplicationList';
 import CssBaseline from '@mui/material/CssBaseline';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
-import PrimarySearchAppBar from './components/SearchBar';
+import PrimarySearchAppBar from './components/PrimarySearchAppBar';
 import { Route, Routes } from "react-router-dom"
 
 const theme = createTheme();
